move Group endpoints from aaa to cms (#442)
